requirements.txt
behave==1.2.6
certifi==2020.12.5
chardet==4.0.0
idna==2.10
parse==1.19.0
parse-type==0.5.2
python-decouple==3.4
requests==2.25.1
selenium==3.141.0
six==1.15.0
urllib3==1.26.4
